requirements.txt
discord.py
aiohttp
asyncpg
uvloop
black
dblpy